requirements.txt
SQLAlchemy~=1.4.29
black
click~=7.1.2
jupyter
linkml
linkml-runtime
pandas_profiling
pandas~=1.3.5
python-dotenv
sklearn